Fixing up more json models
