+++ YoutubeExplode/ChanelHandler/JsonApiModels/CommandMetadata.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace YoutubeExplode.ChanelHandler.JsonApiModels
+{
+    internal class CommandMetadata
+    {
+        [JsonProperty("webCommandMetadata")]
+        public WebCommandMetadata WebCommandMetadata { get; set; }
+    }
+}
