+++ YoutubeExplode/ChanelHandler/JsonApiModels/Info.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace YoutubeExplode.ChanelHandler.JsonApiModels
+{
+    internal class Info
+    {
+        [JsonProperty("st")]
+        public long St { get; set; }
+    }
+}
