+++ YoutubeExplode/ChanelHandler/JsonApiModels/ServiceTrackingParam.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace YoutubeExplode.ChanelHandler.JsonApiModels
+{
+    internal class ServiceTrackingParam
+    {
+        [JsonProperty("service")]
+        public string Service { get; set; }
+
+        [JsonProperty("params")]
+        public List<Param> Params { get; set; }
+    }
+}
