+++ YoutubeExplode/ChanelHandler/JsonApiModels/Timing.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace YoutubeExplode.ChanelHandler.JsonApiModels
+{
+    internal class Timing
+    {
+        [JsonProperty("info")]
+        public Info Info { get; set; }
+    }
+}
