+++ YoutubeExplode/ChanelHandler/JsonApiModels/UrlEndpoint.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace YoutubeExplode.ChanelHandler.JsonApiModels
+{
+    internal class UrlEndpoint
+    {
+        [JsonProperty("url")]
+        public string Url { get; set; }
+    }
+}
