+++ YoutubeExplode/ChanelHandler/JsonApiModels/YtConfigData.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace YoutubeExplode.ChanelHandler.JsonApiModels
+{
+    internal class YtConfigData
+    {
+        [JsonProperty("csn")]
+        public string Csn { get; set; }
+
+        [JsonProperty("visitorData")]
+        public string VisitorData { get; set; }
+    }
+}
